Move reduce tests
