+++ crates/cubek-reduce/tests/it/reduce_shared.rs
