+++ crates/cubek-reduce/tests/it/test_case.rs
