+++ crates/cubek-reduce/tests/lib.rs
@@ -1,2 +1,2 @@
-#[cfg(feature = "extended")]
-pub mod extended;
+// Integration tests
+pub mod it;
